docker/requirements.txt
requests
pyodbc
camunda-external-task-client-python3
pydantic
